// src/components/interface/interface.ts
      export interface Pepole {
        name: string;
        height: string;
        mass: string;
        hair_color: string;
        skin_color: string;
        birth_year: string;
        url:string;
      }
      
      export interface Filme {
        title: string;
        director: string;
        producer: string;
        release_date: string;
        url:string;


      }
      
      export interface Planet {
        name: string;
        rotation_period: string;
        orbital_period: string;
        diameter: string;
        climate: string;
        terrain: string;
        gravity: string;
        url:string;


      }